refactor: rename types and fix type errors with courses
