+++ src/modules/courses/types/courses.ts
@@ -1,13 +1,22 @@
-export enum CourseStatus {
-  ACTIVE = 'Active',
-  INACTIVE = 'Inactive',
-}
+import { CourseStates, CourseModuleTypes } from '@/modules/courses/lib/courses';
 
-export interface CourseProps {
+export type CourseStatus = (typeof CourseStates)[keyof typeof CourseStates];
+export type CourseModuleType =
+  (typeof CourseModuleTypes)[keyof typeof CourseModuleTypes];
+
+export interface CourseModule {
+  id: string;
   title: string;
   description: string;
-  link: string;
+  type: CourseModuleType;
+}
+
+export interface Course {
+  id: string;
+  title: string;
+  description: string;
   imageUrl: string;
   status: CourseStatus;
+  modules: CourseModule[];
   date: Date;
 }
